requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
APScheduler==3.10.3
arcface==0.0.8
array-record==0.4.1
asgiref==3.7.2
astropy==5.3.4
asttokens==2.4.0
astunparse==1.6.3
asyncio==3.4.3
attr==0.3.1
attrs==23.1.0
azure-core==1.29.4
azure-storage-blob==12.18.2
backcall==0.2.0
bbox-visualizer==0.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==5.0.1
blinker==1.6.2
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
deepface==0.0.81
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==3.2.20
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dlib==19.24.2
dm-tree==0.1.8
dnspython==2.4.2
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
etils==1.5.2
executing==1.2.0
expiringdict==1.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.110.0
filelock==3.12.4
Fiona==1.9.4.post1
fire==0.5.0
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.2
gast==0.4.0
gdown==4.7.1
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.37
google-ai-generativelanguage==0.4.0
google-api-core==2.11.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
htmlmin==0.1.12
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.4
ijson==3.2.3
importlib-resources==6.1.1
imutils==0.5.4
inflection==0.5.1
ipykernel==6.26.0
ipython==8.15.0
ipywidgets==8.1.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jax==0.4.16
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==3.2.0
jupyter-bbox-widget==0.5.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyterlab-widgets==3.0.9
keras==2.14.0
kiwisolver==1.4.5
label-studio==1.8.2.post1
label-studio-converter==0.0.54rc0
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
libclang==16.0.6
llvmlite==0.42.0
lockfile==0.12.2
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
mtcnn==0.1.1
nest-asyncio==1.5.8
networkx==3.1
nltk==3.6.7
numba==0.59.0
numpy==1.24.3
oauthlib==3.2.2
openai==1.7.2
openai-whisper==20231117
openapi-codec==1.3.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.3.1
platformdirs==3.10.0
plotly==5.16.1
promise==2.3
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.25.3
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pyerfa==2.0.1.1
Pygments==2.16.1
pylabel==0.1.53
pymongo==4.4.1
pyparsing==3.0.9
pyproj==3.6.0
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-json-logger==2.0.4
python-multipart==0.0.9
python-socketio==5.11.0
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==25.1.1
redbird==0.7.1
redis==3.5.3
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rocketry==2.5.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
schedule==1.2.1
scheduler==0.8.5
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.31.0
setuptools==68.2.2
shapely==2.0.1
simple-websocket==1.0.0
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post9
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.6.0
toml==0.10.2
torch==2.2.0.dev20231001
torchaudio==2.2.0.dev20231001
torchvision==0.17.0.dev20231001
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typeguard==4.1.5
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.18.0
ujson==5.8.0
ultralytics==8.0.190
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==12.0
Werkzeug==2.3.7
wheel==0.41.3
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
xmljson==0.2.0
zipp==3.17.0
